+++ point.h
@@ -5,7 +5,7 @@ class Point
 {
 public:
     Point();
-    Point(double,double, int);
+    Point(double,double, double);
     double x;
     double y;
     int Class;
